# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Restaurant Platform API
  version: 1.0.0
  description: >
    REST API para el sistema de gestión de pedidos, reseñas y analítica de
    restaurantes sobre MongoDB Atlas (CC3089 Bases de Datos 2).


    **Recursos principales**: Users, Restaurants, Menu Items, Carts, Orders,
    Reviews, Files (GridFS), Analytics.
servers:
  - url: http://localhost:3000
    description: Servidor local de desarrollo
...
paths:
...
  /users:
    post:
      summary: Crear usuario
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, name]
              properties:
                email:
                  type: string
                  example: demo@example.com
                name:
                  type: string
                  example: Usuario Demo
                phone:
                  type: string
                  example: "+502 5555"
                role:
                  type: string
                  enum: [customer, restaurant_admin]
                  default: customer
                defaultAddress:
                  $ref: '#/components/schemas/Address'
      responses:
        '201':
          description: Usuario creado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '409':
          description: Email ya registrado.

    get:
      summary: Listar usuarios
      tags: [Users]
      parameters:
        - in: query
          name: role
          schema:
            type: string
            enum: [customer, restaurant_admin]
        - $ref: '#/components/parameters/skip'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: Lista paginada de usuarios.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
  /users/{id}:
    get:
      summary: Obtener usuario por ID
      tags: [Users]
      parameters:
        - $ref: '#/components/parameters/objectId'
      responses:
        '200':
          description: Datos del usuario.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: Usuario no encontrado.

    patch:
      summary: Actualizar perfil de usuario
      tags: [Users]
      parameters:
        - $ref: '#/components/parameters/objectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                phone:
                  type: string
                defaultAddress:
                  $ref: '#/components/schemas/Address'
      responses:
        '200':
          description: Resultado de la actualización.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateResult'
        '404':
          description: Usuario no encontrado.

    delete:
      summary: Eliminar usuario
      tags: [Users]
      parameters:
        - $ref: '#/components/parameters/objectId'
      responses:
        '200':
          description: Usuario eliminado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResult'
        '404':
          description: Usuario no encontrado.
  /users/{id}/favorites:
    patch:
      summary: Agregar restaurante a favoritos ($push + $slice Subset Pattern)
      tags: [Users]
      parameters:
        - $ref: '#/components/parameters/objectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [restaurantId]
              properties:
                restaurantId:
                  type: string
                  example: 65f1c5b8e8a1b2c3d4e5f678
      responses:
        '200':
          description: Favorito agregado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateResult'
...
  /reviews:
    post:
      summary: Crear reseña (referenciada a orden delivered)
      description: >
        Valida que la orden existe, pertenece al usuario y tiene status delivered.
        Impide reseñas duplicadas por orden.
      tags: [Reviews]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, orderId, restaurantId, rating]
              properties:
                userId:
                  type: string
                orderId:
                  type: string
                restaurantId:
                  type: string
                rating:
                  type: integer
                  minimum: 1
                  maximum: 5
                title:
                  type: string
                  example: Excelente servicio
                comment:
                  type: string
                tags:
                  type: array
                  items:
                    type: string
                  example: [recomendado, buen-servicio]
      responses:
        '201':
          description: Reseña creada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '400':
          description: Orden no entregada o no pertenece al usuario.
        '409':
          description: Ya existe reseña para esta orden.

    get:
      summary: Listar reseñas (filtros, text search, paginación)
      tags: [Reviews]
      parameters:
        - in: query
          name: restaurantId
          schema:
            type: string
        - in: query
          name: userId
          schema:
            type: string
        - in: query
          name: rating
          schema:
            type: integer
            minimum: 1
            maximum: 5
        - in: query
          name: tag
          schema:
            type: string
          description: Filtrar por tag (multikey index).
        - in: query
          name: q
          schema:
            type: string
          description: Búsqueda full-text ($text) en title y comment.
        - $ref: '#/components/parameters/skip'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: Lista de reseñas.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Review'

    delete:
      summary: Eliminar reseñas por filtro (deleteMany)
      tags: [Reviews]
      parameters:
        - in: query
          name: restaurantId
          schema:
            type: string
        - in: query
          name: userId
          schema:
            type: string
        - in: query
          name: before
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Reseñas eliminadas.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResult'
        '400':
          description: Al menos un filtro requerido.
  /reviews/{id}:
    get:
      summary: Obtener reseña por ID
      tags: [Reviews]
      parameters:
        - $ref: '#/components/parameters/objectId'
      responses:
        '200':
          description: Detalle de la reseña.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '404':
          description: Reseña no encontrada.

    delete:
      summary: Eliminar reseña
      tags: [Reviews]
      parameters:
        - $ref: '#/components/parameters/objectId'
      responses:
        '200':
          description: Reseña eliminada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResult'
        '404':
          description: Reseña no encontrada.
  /reviews/{id}/response:
    patch:
      summary: Respuesta del restaurante a una reseña (embebido 1:1)
      tags: [Reviews]
      parameters:
        - $ref: '#/components/parameters/objectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message:
                  type: string
                  example: Gracias por tu reseña. Esperamos verte pronto.
      responses:
        '200':
          description: Respuesta registrada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateResult'
        '404':
          description: Reseña no encontrada.
  /reviews/{id}/tag:
    patch:
      summary: Agregar tag a reseña ($addToSet)
      tags: [Reviews]
      parameters:
        - $ref: '#/components/parameters/objectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tag]
              properties:
                tag:
                  type: string
                  example: recomendado
      responses:
        '200':
          description: Tag agregado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateResult'
  /reviews/{id}/helpful:
    patch:
      summary: Votar reseña como útil ($addToSet)
      tags: [Reviews]
      parameters:
        - $ref: '#/components/parameters/objectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [voterId]
              properties:
                voterId:
                  type: string
                  description: ObjectId del usuario que vota.
      responses:
        '200':
          description: Voto registrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateResult'
...
components:
  parameters:
    objectId:
      in: path
      name: id
      required: true
      schema:
        type: string
      description: ObjectId de MongoDB (24 caracteres hex).
    skip:
      in: query
      name: skip
      schema:
        type: integer
        minimum: 0
        default: 0
    limit:
      in: query
      name: limit
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
    GeoPoint:
      type: object
      properties:
        type:
          type: string
          example: Point
        coordinates:
          type: array
          items:
            type: number
          example: [-90.5069, 14.5943]
          description: "[longitude, latitude]"
    Address:
      type: object
      properties:
        street:
          type: string
          example: 7a Avenida 12-30
        city:
          type: string
          example: Guatemala
        zone:
          type: string
          example: Zona 9
        coordinates:
          $ref: '#/components/schemas/GeoPoint'
    User:
      type: object
      properties:
        _id:
          type: string
        email:
          type: string
        name:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [customer, restaurant_admin]
        defaultAddress:
          $ref: '#/components/schemas/Address'
        orderHistory:
          type: array
          items:
            type: string
        favoriteRestaurants:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
...
    Review:
      type: object
      properties:
        _id:
          type: string
        userId:
          type: string
        restaurantId:
          type: string
        orderId:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        title:
          type: string
        comment:
          type: string
        tags:
          type: array
          items:
            type: string
        restaurantResponse:
          type: object
          nullable: true
          properties:
            message:
              type: string
            respondedAt:
              type: string
              format: date-time
        helpfulVotes:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
    UpdateResult:
      type: object
      properties:
        updated:
          type: integer
          example: 1
    DeleteResult:
      type: object
      properties:
        deleted:
          type: integer
          example: 1
...